Move CompassFragment into compass package
