+++ app/src/main/java/de/haw_hamburg/sensorapp/compass/Compass.java
@@ -1,4 +1,4 @@
-package de.haw_hamburg.sensorapp;
+package de.haw_hamburg.sensorapp.compass;
 
 import android.content.Context;
 import android.hardware.Sensor;
